Assignment 9/10 - Fixed File Structure (Hopefully)
