+++ Assignment.Tests/Assignment.Tests.csproj
@@ -2,6 +2,8 @@
 
   <PropertyGroup>
     <IsPackable>false</IsPackable>
+	<TargetFramework>.net8.0</TargetFramework>
+	<LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
@@ -9,7 +11,7 @@
     <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
     <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
     <PackageReference Include="coverlet.collector" Version="1.0.1" />
-    <ProjectReference Include="..\Assignment\Assignment.csproj" />
+    <ProjectReference Include="..\Assignment\Assignment\Assignment.csproj" />
 	  <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
   </ItemGroup>
