+++ Assignment.sln
@@ -5,12 +5,12 @@ VisualStudioVersion = 17.9.34607.119
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{E316FCD7-C0C8-4F2F-87FB-F92921756F02}"
 	ProjectSection(SolutionItems) = preProject
-		Directory.Build.props = Directory.Build.props
+		Assignment\Assignment\Directory.Build.props = Assignment\Assignment\Directory.Build.props
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment", "Assignment\Assignment\Assignment.csproj", "{0030CB91-2D00-402A-ABBA-1235F27FBE38}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment.Tests", "Assignment\Assignment.Tests\Assignment.Tests.csproj", "{1B47DAF8-FDEB-48A6-9FF5-2B5076C7F1BA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment.Tests", "Assignment.Tests\Assignment.Tests.csproj", "{C1126C49-15F6-49D4-981C-520FD7080608}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,10 +22,10 @@ Global
 		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1B47DAF8-FDEB-48A6-9FF5-2B5076C7F1BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1B47DAF8-FDEB-48A6-9FF5-2B5076C7F1BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1B47DAF8-FDEB-48A6-9FF5-2B5076C7F1BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1B47DAF8-FDEB-48A6-9FF5-2B5076C7F1BA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1126C49-15F6-49D4-981C-520FD7080608}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1126C49-15F6-49D4-981C-520FD7080608}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1126C49-15F6-49D4-981C-520FD7080608}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1126C49-15F6-49D4-981C-520FD7080608}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
